Homework_12_Maven + Java Stream API.
In this task, the Gson dependency was added to the project to facilitate working with JSON files. Two key classes were implemented: UserLoader, which utilizes Gson to load users from a JSON file, and UserProcessor, which filters and sorts the resulting user list. This setup efficiently handles the processing and analysis of user data within the project.